requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.10
botocore==1.27.10
Django==4.0.4
django-debug-toolbar==3.5.0
django-environ==0.8.1
django-s3-sqlite==0.0.5
jmespath==1.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.6.0
serverless-wsgi==3.0.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
